Add check for server.sh process before starting it -----------------------------------------------
This commit introduces an if statement to verify whether the server.sh process is already running before attempting to start it again. If the process is found to be running, an informational message is
printed. Otherwise, tsp is used to stop and start the server.sh script with a delay of 2 seconds.
	modified:   ollama.sh