Removing unnecessary `set -e` for better error handling flexibility in shell/meta.sh
This commit message indicates that the line containing `set -e` has been removed from the file
`shell/meta.sh`. The reason for this change is to improve the error handling flexibility by not
automatically exiting on errors. This can be useful in certain situations where it is desirable
to allow scripts to continue running even when errors occur.

	modified:   meta.sh